requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-environ==0.10.0
argon2-cffi==21.3.0
cryptography==39.0.1
pycryptodome==3.17.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0